Refactor test base
